requirements.txt
line-bot-sdk==3.11.0
django
djangorestframework
drf-yasg

ultralytics
roboflow
django-allauth

# Dependencies of ultralytics
# pillow
# numpy==1.26.4
# torch==2.2.2
# opencv-python==4.10.0